@@ -11,7 +11,7 @@
   "scripts": {
     "dev": "cross-env NODE_ENV=development webpack-dev-server --open --hot",
     "build": "rm -rf ./dist && bili src/vue-thailand-address-autocomplete.vue",
-    "build:demo": "cross-env NODE_ENV=production webpack --progress --hide-modules"
+    "build:demo": "cross-env NODE_ENV=production webpack --progress"
   },
   "repository": {
     "type": "git",
@@ -20,7 +20,7 @@
   "dependencies": {
     "thai-address-database": "^0.0.27",
     "v-click-outside": "^1.0.4",
-    "vue": "^2.5.11"
+    "vue": "^3.2.31"
   },
   "browserslist": [
     "> 1%",
@@ -36,10 +36,12 @@
     "cross-env": "^5.0.5",
     "css-loader": "^0.28.7",
     "file-loader": "^1.1.4",
-    "rollup-plugin-vue": "^3.0.0",
-    "vue-loader": "^13.0.5",
-    "vue-template-compiler": "^2.4.4",
-    "webpack": "^3.6.0",
-    "webpack-dev-server": "^2.9.1"
+    "rollup-plugin-vue": "^6.0.0-beta.10",
+    "vue-loader": "^17.0.0",
+    "vue-style-loader": "^4.1.3",
+    "vue-template-compiler": "^2.7.10",
+    "webpack": "^4.1.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^4.11.0"
   }
 }
